Updated tests, and fixed some bugs
