+++ src/parse.jl
@@ -4,10 +4,10 @@ import ArgParse, Git, FileIO, JLD2
 import ArgParse.@add_arg_table
 import ArgParse.ArgParseSettings
 
-export
-    parse_args,
-    @add_arg_table,
-    ArgParseSettings
+# export
+#     parse_args,
+#     @add_arg_table,
+#     ArgParseSettings
 
 
 # HASHER(x) = hash(x)
